refactor: move organizations folder to admin-console, refs AC-1202
