+++ apps/web/src/app/admin-console/organizations/members/components/member-dialog/member-dialog.component.ts
@@ -22,7 +22,7 @@ import {
   GroupView,
   OrganizationUserAdminView,
   UserAdminService,
-} from "../../../../../organizations/core";
+} from "../../../core";
 import {
   AccessItemType,
   AccessItemValue,
