+++ apps/web/src/app/admin-console/organizations/members/people.component.ts
@@ -55,13 +55,13 @@ import {
 
 import { EntityEventsComponent } from "../../../admin-console/organizations/manage/entity-events.component";
 import { BasePeopleComponent } from "../../../common/base.people.component";
-import { GroupService } from "../../../organizations/core";
-import { OrganizationUserView } from "../../../organizations/core/views/organization-user.view";
-import { BulkConfirmComponent } from "../../../organizations/members/components/bulk/bulk-confirm.component";
-import { BulkRemoveComponent } from "../../../organizations/members/components/bulk/bulk-remove.component";
-import { BulkRestoreRevokeComponent } from "../../../organizations/members/components/bulk/bulk-restore-revoke.component";
-import { BulkStatusComponent } from "../../../organizations/members/components/bulk/bulk-status.component";
+import { GroupService } from "../core";
+import { OrganizationUserView } from "../core/views/organization-user.view";
 
+import { BulkConfirmComponent } from "./components/bulk/bulk-confirm.component";
+import { BulkRemoveComponent } from "./components/bulk/bulk-remove.component";
+import { BulkRestoreRevokeComponent } from "./components/bulk/bulk-restore-revoke.component";
+import { BulkStatusComponent } from "./components/bulk/bulk-status.component";
 import {
   MemberDialogResult,
   MemberDialogTab,
