+++ apps/web/src/app/admin-console/organizations/organization.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from "@angular/core";
 
-import { CoreOrganizationModule } from "../../organizations/core";
 import { SharedOrganizationModule } from "../../organizations/shared";
 
+import { CoreOrganizationModule } from "./core";
 import { GroupAddEditComponent } from "./manage/group-add-edit.component";
 import { GroupsComponent } from "./manage/groups.component";
 import { OrganizationsRoutingModule } from "./organization-routing.module";
