+++ apps/web/src/app/admin-console/organizations/shared/shared-organization.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from "@angular/core";
 
-import { AccessSelectorModule } from "../../admin-console/organizations/shared/components/access-selector/access-selector.module";
-import { CollectionDialogModule } from "../../admin-console/organizations/shared/components/collection-dialog";
-import { SharedModule } from "../../shared/shared.module";
+import { SharedModule } from "../../../shared/shared.module";
 
+import { AccessSelectorModule } from "./components/access-selector/access-selector.module";
+import { CollectionDialogModule } from "./components/collection-dialog";
 import { SearchInputComponent } from "./components/search-input/search-input.component";
 
 @NgModule({
