+++ apps/web/src/app/organizations/shared/index.ts
@@ -1,2 +1,2 @@
-export * from "./shared-organization.module";
+export * from "../../admin-console/organizations/shared/shared-organization.module";
 export * from "../../admin-console/organizations/shared/components/collection-dialog";
