+++ apps/web/src/app/vault/individual-vault/vault-filter/shared/models/vault-filter.type.ts
@@ -3,7 +3,7 @@ import { ITreeNodeObject } from "@bitwarden/common/src/models/domain/tree-node";
 import { FolderView } from "@bitwarden/common/src/vault/models/view/folder.view";
 import { CipherType } from "@bitwarden/common/vault/enums/cipher-type";
 
-import { CollectionAdminView } from "../../../../../organizations/core";
+import { CollectionAdminView } from "../../../../../admin-console/organizations/core";
 
 export type CipherStatus = "all" | "favorites" | "trash" | CipherType;
