+++ apps/web/src/app/vault/org-vault/vault-filter/vault-filter.service.ts
@@ -14,8 +14,8 @@ import { TreeNode } from "@bitwarden/common/models/domain/tree-node";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { FolderService } from "@bitwarden/common/vault/abstractions/folder/folder.service.abstraction";
 
-import { CollectionAdminView } from "../../../organizations/core";
-import { CollectionAdminService } from "../../../organizations/core/services/collection-admin.service";
+import { CollectionAdminView } from "../../../admin-console/organizations/core";
+import { CollectionAdminService } from "../../../admin-console/organizations/core/services/collection-admin.service";
 import { VaultFilterService as BaseVaultFilterService } from "../../individual-vault/vault-filter/services/vault-filter.service";
 import { CollectionFilter } from "../../individual-vault/vault-filter/shared/models/vault-filter.type";
