+++ bitwarden_license/bit-web/src/app/admin-console/providers/manage/bulk/bulk-remove.component.ts
@@ -1,11 +1,11 @@
 import { Component, Input } from "@angular/core";
 
 import { ProviderUserBulkRequest } from "@bitwarden/common/admin-console/models/request/provider/provider-user-bulk.request";
-import { BulkRemoveComponent as OrganizationBulkRemoveComponent } from "@bitwarden/web-vault/app/organizations/members/components/bulk/bulk-remove.component";
+import { BulkRemoveComponent as OrganizationBulkRemoveComponent } from "@bitwarden/web-vault/app/admin-console/organizations/members/components/bulk/bulk-remove.component";
 
 @Component({
   templateUrl:
-    "../../../../../../../../apps/web/src/app/organizations/members/components/bulk/bulk-remove.component.html",
+    "../../../../../../../../apps/web/src/app/admin-console/organizations/members/components/bulk/bulk-remove.component.html",
 })
 export class BulkRemoveComponent extends OrganizationBulkRemoveComponent {
   @Input() providerId: string;
